Update for DataTransactionBuilder changes. Add AbstractSingleTargetDualProcessor to implement both a DataProcessor and ValueProcessor. Implement ExpirableData for Endermites.
