+++ src/main/java/org/spongepowered/common/data/manipulator/immutable/entity/ImmutableSpongeExpirableData.java
@@ -38,7 +38,7 @@ public class ImmutableSpongeExpirableData extends AbstractImmutableBoundedCompar
     implements ImmutableExpirableData{
 
     public ImmutableSpongeExpirableData(int value, int upperBound) {
-        super(ImmutableExpirableData.class, value, Keys.EXPIRATION_TICKS, intComparator(), SpongeExpirableData.class, 0, upperBound, 0);
+        super(ImmutableExpirableData.class, value, Keys.EXPIRATION_TICKS, intComparator(), SpongeExpirableData.class, 0, upperBound, upperBound);
     }
 
     @Override
