+++ src/main/java/org/spongepowered/common/data/processor/value/entity/RespawnLocationValueProcessor.java
@@ -26,7 +26,6 @@ package org.spongepowered.common.data.processor.value.entity;
 
 import com.flowpowered.math.vector.Vector3d;
 import com.google.common.collect.ImmutableMap;
-import org.spongepowered.api.data.DataTransactionBuilder;
 import org.spongepowered.api.data.DataTransactionResult;
 import org.spongepowered.api.data.key.Keys;
 import org.spongepowered.api.data.value.ValueContainer;
@@ -53,12 +52,13 @@ public class RespawnLocationValueProcessor extends AbstractSpongeValueProcessor<
         if (container instanceof ISpongeUser) {
             ImmutableMap<UUID, Vector3d> removed = ((ISpongeUser) container).removeAllBeds();
             if (!removed.isEmpty()) {
-                return DataTransactionBuilder.builder().result(DataTransactionResult.Type.SUCCESS).replace(constructImmutableValue(removed)).build();
+                return DataTransactionResult
+                    .builder().result(DataTransactionResult.Type.SUCCESS).replace(constructImmutableValue(removed)).build();
             } else {
-                return DataTransactionBuilder.successNoData();
+                return DataTransactionResult.successNoData();
             }
         }
-        return DataTransactionBuilder.failNoData();
+        return DataTransactionResult.failNoData();
     }
 
     @Override
