+++ src/main/resources/common_at.cfg
@@ -119,7 +119,6 @@ public-f net.minecraft.entity.EntityLivingBase field_94063_bt # _combatTracker
 public net.minecraft.entity.EntityTracker field_72794_c # trackedEntityHashTable
 public net.minecraft.entity.monster.EntitySlime func_70799_a(I)V # setSlimeSize
 public net.minecraft.entity.monster.EntityEndermite field_175497_b # lifetime
-public net.minecraft.entity.monster.EntityPigZombie field_70837_d # angerLevel
 public net.minecraft.entity.item.EntityItem field_70292_b # age
 public net.minecraft.entity.passive.EntityWolf field_70925_g # isWet
 public net.minecraft.entity.passive.EntityWolf field_70928_h # isShaking
